Make ClientOpts use a password converter that respects the env:PASSWORD
